Extracts hot reload utilities, makes placeholderText contribution hot reloadable
